requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
aiomysql
jinja2
python-jose
python-multipart
passlib
bcrypt
PyJWT
pydantic
bcrypt